fix on creating checkout object
